// generated/orval/fetch/google-health-api/models/foodAccessLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Required. The access level of the food.
 */
export type FoodAccessLevel = typeof FoodAccessLevel[keyof typeof FoodAccessLevel];


export const FoodAccessLevel = {
  FOOD_ACCESS_LEVEL_UNSPECIFIED: 'FOOD_ACCESS_LEVEL_UNSPECIFIED',
  FOOD_ACCESS_LEVEL_PUBLIC: 'FOOD_ACCESS_LEVEL_PUBLIC',
  FOOD_ACCESS_LEVEL_PRIVATE: 'FOOD_ACCESS_LEVEL_PRIVATE',
} as const;